Refactor: move legacy effective_date_transaction into a separate file
