+++ .gitignore
@@ -1,3 +1,6 @@
+# Editor
+*~
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
